class attributed updated LegalPage. index.js updated to React 18 format.
React does not recognise the "class" attribute. "class" attributes on the legal page were updated to "className" (was mistakenly left out of the prior similar update for the other pages). 
index.js was using render instead of createRoot, which forced React to run in the older React 17 version. This was fixed so that react now runs as version 18.