Fixed bugs in form. Set up formCheck on Photobooth page.
Fixed bugs in form where the values were not being passed to the regex check functions.
Set up a check where whether the form has been completed by the user is stored and passed to the Photobooth component which shows a message prompting the user to complete the form if they have not done so as yet.